Patch support for check mark (resolve #18)
As tested, the current mocha and mocha-chrome libraries output garbled output for passing unit tests, resulting in `âœ“` instead of a check mark.  

I assume the issue is probably related to content-encoding. Something is not being treated as UTF-8. 

In the meantime, this patch resolves the most common issue (the broken checkmark), in a way that (based on Mocha's source code) should print a compatible character (check mark on Linux/macOS, square root on Windows).

Thank you :smile: